// src/profile/dto/profile.dto.ts
export class ProfileDto{
    email: string;
    username: string;
    username_tag: string;
    about: string;
    interest: string;    
    display_name: string;    
    gender: string;    
    birthday: Date;    
    height_unit: string;    
    height_value: Number;    
    weight_unit: string;
    weight_value: Number;
    zodiac: string;
    horoscope: string;
    profile_pic: string;
    profile_pic_type: string;
}